Refactor import paths for PrismaService in service tests
- Updated import paths for PrismaService in both PostsService and UsersService test files to reflect the new structure, enhancing code clarity and maintainability.
